Struct guardhaus::authentication_info::AuthenticationInfo

pub struct AuthenticationInfo {
    pub digest: Option<String>,
    pub next_nonce: Option<String>,
    pub qop: Option<Qop>,
    pub client_nonce: Option<String>,
    pub nonce_count: Option<NonceCount>,
}

Parameters for the Authentication-Info header.

Fields
digest: Option<String>

The digest of the entity body, parameter name digest in RFC 2069, rspauth otherwise

next_nonce: Option<String>

nextnonce - per RFC 7616, "the nonce the server wishes the client to use for a future authentication response."

qop: Option<Qop>

Quality of protection

client_nonce: Option<String>

Cryptographic nonce from the client

nonce_count: Option<NonceCount>

Nonce count, parameter name nc