Struct guardhaus::digest::Digest

pub struct Digest {
    pub username: Username,
    pub realm: String,
    pub nonce: String,
    pub nonce_count: Option<NonceCount>,
    pub response: String,
    pub request_uri: String,
    pub algorithm: HashAlgorithm,
    pub qop: Option<Qop>,
    pub client_nonce: Option<String>,
    pub opaque: Option<String>,
    pub charset: Option<Charset>,
    pub userhash: bool,
}

Parameters for the Authorization header when using the Digest scheme.

The parameters are described in more detail in RFC 7616. Unless otherwise noted, the parameter name maps to the struct variable name.

Fields
username: Username

Either the user name or the user hash (if userhash is true - see RFC 7616, section 3.4.4).

realm: String

Authentication realm.

nonce: String

Cryptographic nonce.

nonce_count: Option<NonceCount>

Nonce count, parameter name nc. Optional only in RFC 2067 mode.

response: String

The hexadecimal digest of the payload as described by the RFCs.

request_uri: String

Either the absolute path or URI of the HTTP request, parameter name uri.

algorithm: HashAlgorithm

The hash algorithm to use when generating the response.

qop: Option<Qop>

Quality of protection. Optional only in RFC 2067 mode.

client_nonce: Option<String>

Cryptographic nonce from the client. Optional only in RFC 2067 mode.

opaque: Option<String>

Optional opaque string.

charset: Option<Charset>

The character set to use when generating the A1 value or the userhash. Added for RFC 7616.

userhash: bool

Whether username is a userhash. Added for RFC 7616.

Methods

impl Digest

fn userhash(
algorithm: &HashAlgorithm,
username: Vec<u8>,
realm: String
) -> String

Generates a userhash, as defined in RFC 7616, section 3.4.4.

fn validate_userhash(&self, username: Username) -> bool

Validates a userhash (as defined in RFC 7616, section 3.4.4), given a Digest header.

If userhash is false, returns false.

fn simple_hashed_a1(
algorithm: &HashAlgorithm,
username: Username,
realm: String,
password: String
) -> String

Generates a simple hexadecimal digest from an A1 value and given algorithm.

This is intended to be used in applications that use the htdigest style of secret hash generation.

To see how a simple A1 value is constructed, see RFC 7616, section 3.4.2. This is the definition when the algorithm is "unspecified".

fn using_password(
&self,
method: Method,
entity_body: &[u8],
password: String
) -> Result<String, Error>

Generates a digest, given an HTTP request and a password.

entity_body is defined in RFC 2616, secion 7.2.

fn using_hashed_a1(
&self,
method: Method,
entity_body: &[u8],
a1: String
) -> Result<String, Error>

Generates a digest, given an HTTP request and a hexadecimal digest of an A1 string.

entity_body is defined in RFC 2616, secion 7.2.

This is intended to be used in applications that use the htdigest style of secret hash generation.

fn validate_using_password(
&self,
method: Method,
entity_body: &[u8],
password: String
) -> bool

Validates a Digest.response, given an HTTP request and a password.

entity_body is defined in RFC 2616, secion 7.2.

fn validate_using_userhash_and_password(
&self,
method: Method,
entity_body: &[u8],
username: Username,
password: String
) -> bool

Validates a Digest.username and Digest.response, given an HTTP request, a username, and a password. If a userhash is specified, that is validated first.

entity_body is defined in RFC 2616, secion 7.2.

fn validate_using_hashed_a1(
&self,
method: Method,
entity_body: &[u8],
a1: String
) -> bool

Validates a Digest.response, given an HTTP request and a hexadecimal digest of an A1 string.

entity_body is defined in RFC 2616, secion 7.2.

This is intended to be used in applications that use the htdigest style of secret hash generation.