Enum guardhaus::types::HashAlgorithm

pub enum HashAlgorithm {
    MD5,
    MD5Session,
    SHA256,
    SHA256Session,
    SHA512256,
    SHA512256Session,
}

Allowable hash algorithms for the algorithm parameter.

Variants
MD5

MD5

MD5Session

MD5-sess

SHA256

SHA-256

SHA256Session

SHA-256-sess

SHA512256

SHA-512-256

SHA512256Session

SHA-512-256-sess

Methods

impl HashAlgorithm

fn hex_digest(&self, data: &[u8]) -> String

Generate a hexadecimal representation of the output of a cryptographic hash function, given data and the algorithm.